BtBRQueue: set timeout on waiting for available data during disconnection
The latch may cause the main loop thread to hang when a device is
getting disconnected from the UI and eventually cause Android to restart
the app altogether. Setting a time out in the await call will cause the
latch to count down on itself after the timeout so that we do not hang
for too long.